[GB-77] doc : 폴더명 변경 등
