+++ src/hooks/useDashboardStats.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import axiosInstance from '../API/axiosInstance';
+import axiosInstance from '../api/axiosInstance';
 
 interface StatBoxProps {
   label: string;
