+++ src/pages/AdminSetting.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useMemo, useEffect } from 'react';
 import {
   Table,
   Thead,
@@ -6,7 +6,6 @@ import {
   Tr,
   Th,
   Td,
-  TableCaption,
   TableContainer,
   Button,
   Checkbox,
@@ -19,29 +18,15 @@ import {
 } from '@chakra-ui/react';
 import { ChevronDownIcon, ChevronUpIcon } from '@chakra-ui/icons';
 import ConfirmModal from '../Components/common/ConfirmModal';
-import axiosInstance from '../API/axiosInstance';
 import Pagination from '../Components/common/Pagination';
-
-interface User {
-  userId: number;
-  email: string;
-  nickname: string;
-  birthYear: number;
-  age: number;
-  gender: string;
-  location: string;
-  active: string;
-}
+import axiosInstance from '../api/axiosInstance';
 
 interface DataItem {
-  id: number;
-  nickname: string;
-  email: string;
-  birthYear: number;
-  age: number;
-  gender: string;
-  location: string;
-  active: string;
+  adminId: number;
+  adminEmail: string;
+  adminActive: string;
+  createdAt: string;
+  updatedAt: string;
 }
 
 interface SortConfig {
@@ -49,63 +34,30 @@ interface SortConfig {
   direction: 'ascending' | 'descending';
 }
 
-const UserSetting: React.FC = () => {
+const AdminSetting = () => {
   const [data, setData] = useState<DataItem[]>([]);
+
   const [sortConfig, setSortConfig] = useState<SortConfig | null>(null);
   const [selectedRows, setSelectedRows] = useState<number[]>([]);
-  const [searchTerm, setSearchTerm] = useState<string>('');
+  const [searchTerm, setSearchTerm] = useState('');
   const { isOpen, onOpen, onClose } = useDisclosure();
   const [deleteRowIds, setDeleteRowIds] = useState<number[]>([]);
+  const currentUserCanChangePermissions = true; // 현재 유저의 권한 변경 가능 여부 (예시)
 
   // 페이지네이션 상태
   const [currentPage, setCurrentPage] = useState(1);
   const [totalPages, setTotalPages] = useState(1);
   const itemsPerPage = 10;
 
-  // 유저 데이터 조회
+  // 관리자 회원 목록 조회
   useEffect(() => {
-    getUserData();
-  }, [currentPage, sortConfig, itemsPerPage]);
-
-  const getUserData = async () => {
-    const sortKey = sortConfig?.key ?? 'userId';
-    const sortDirection = sortConfig?.direction ?? 'ascending';
-
-    try {
-      const res = await axiosInstance.get<{
-        sort: string;
-        pageNumber: number;
-        pageSize: number;
-        totalPage: number;
-        totalCount: number;
-        users: User[];
-      }>('/admin/user/', {
-        params: {
-          sort: `${sortKey},${sortDirection}`,
-          pageNumber: currentPage - 1,
-          pageSize: itemsPerPage,
-        },
-      });
-
-      const users = res.data.users.map(user => ({
-        id: user.userId,
-        nickname: user.nickname,
-        email: user.email,
-        birthYear: user.birthYear,
-        age: user.age,
-        gender: user.gender,
-        location: user.location,
-        active: user.active,
-      }));
-
-      setData(users);
+    axiosInstance.get('/admin/').then(res => {
+      setData(res.data.admins);
+      console.log(res.data);
       setTotalPages(res.data.totalPage);
-    } catch (error) {
-      console.error('Error fetching user data:', error);
-    }
-  };
+    });
+  }, []);
 
-  // 정렬 기능
   const requestSort = (key: keyof DataItem) => {
     let direction: 'ascending' | 'descending' = 'ascending';
     if (
@@ -129,12 +81,50 @@ const UserSetting: React.FC = () => {
     return <ChevronDownIcon />;
   };
 
-  // 선택 삭제 기능
+  // 관리자 이메일 검색 기능
+  const filteredData = useMemo(() => {
+    return data.filter(item =>
+      item.adminEmail.toLowerCase().includes(searchTerm.toLowerCase()),
+    );
+  }, [data, searchTerm]);
+
+  // 정렬 기능
+  const sortedData = useMemo(() => {
+    const sortableItems = [...filteredData];
+    if (sortConfig !== null) {
+      const { key, direction } = sortConfig;
+
+      sortableItems.sort((a, b) => {
+        const aValue = a[key];
+        const bValue = b[key];
+
+        if (aValue === null && bValue !== null) {
+          return direction === 'ascending' ? -1 : 1;
+        }
+        if (aValue !== null && bValue === null) {
+          return direction === 'ascending' ? 1 : -1;
+        }
+        if (aValue === null && bValue === null) {
+          return 0;
+        }
+
+        if (aValue! < bValue!) {
+          return direction === 'ascending' ? -1 : 1;
+        }
+        if (aValue! > bValue!) {
+          return direction === 'ascending' ? 1 : -1;
+        }
+        return 0;
+      });
+    }
+    return sortableItems;
+  }, [filteredData, sortConfig]);
+
   const handleSelectAll = () => {
     if (selectedRows.length === data.length) {
       setSelectedRows([]);
     } else {
-      setSelectedRows(data.map(item => item.id));
+      setSelectedRows(data.map(item => item.adminId));
     }
   };
 
@@ -151,29 +141,35 @@ const UserSetting: React.FC = () => {
 
   const handleBulkDelete = () => {
     const deletableRows = selectedRows.filter(
-      id => data.find(row => row.id === id)?.active !== 'inactive',
+      id => !(data.find(row => row.adminId === id)?.adminActive === '탈퇴'),
     );
     setDeleteRowIds(deletableRows);
     onOpen();
   };
 
-  const confirmDelete = async () => {
-    console.log('confirmDelete user:', ...deleteRowIds);
-    try {
-      // 각 사용자에 대해 삭제 요청을 서버로 보냄
-      for (const id of deleteRowIds) {
-        console.log(`/admin/user/${id}`);
-        await axiosInstance.delete(`/admin/user/${id}`);
-      }
-      console.log('Users deleted successfully');
-    } catch (error) {
-      console.error('Error deleting users:', error);
-    }
-
-    // 상태 업데이트
-    setData(prevData => prevData.filter(row => !deleteRowIds.includes(row.id)));
+  const confirmDelete = () => {
+    console.log('이 회원을 삭제하시겠습니까?', deleteRowIds);
+    axiosInstance
+      .delete(`/admin/delete/${deleteRowIds}`)
+      .then(res => {
+        console.log(res.data);
+      })
+      .catch(err => {
+        console.error(err);
+      });
+    // setData(prevData =>
+    //   prevData.filter(row => !deleteRowIds.includes(row.adminId)),
+    // );
     setSelectedRows([]);
     onClose();
+  };
+
+  const changePermission = (id: number, newPermission: string) => {
+    setData(prevData =>
+      prevData.map(row =>
+        row.adminId === id ? { ...row, 권한: newPermission } : row,
+      ),
+    );
   };
 
   // 페이지네이션 함수
@@ -186,7 +182,7 @@ const UserSetting: React.FC = () => {
       <VStack spacing={4} align="stretch">
         <HStack spacing={4}>
           <Input
-            placeholder="유저명 또는 이메일 검색"
+            placeholder="관리자명 또는 이메일 검색"
             value={searchTerm}
             onChange={e => setSearchTerm(e.target.value)}
           />
@@ -210,19 +206,9 @@ const UserSetting: React.FC = () => {
                 <Th textAlign="center" fontWeight="bold" fontSize="1rem">
                   ID
                   <IconButton
-                    icon={renderSortIcon('id')}
-                    onClick={() => requestSort('id')}
+                    icon={renderSortIcon('adminId')}
+                    onClick={() => requestSort('adminId')}
                     aria-label="Sort ID"
-                    size="xs"
-                    ml={2}
-                  />
-                </Th>
-                <Th textAlign="center" fontWeight="bold" fontSize="1rem">
-                  유저명
-                  <IconButton
-                    icon={renderSortIcon('nickname')}
-                    onClick={() => requestSort('nickname')}
-                    aria-label="Sort 유저명"
                     size="xs"
                     ml={2}
                   />
@@ -230,63 +216,53 @@ const UserSetting: React.FC = () => {
                 <Th textAlign="center" fontWeight="bold" fontSize="1rem">
                   이메일
                   <IconButton
-                    icon={renderSortIcon('email')}
-                    onClick={() => requestSort('email')}
+                    icon={renderSortIcon('adminEmail')}
+                    onClick={() => requestSort('adminEmail')}
                     aria-label="Sort 이메일"
                     size="xs"
                     ml={2}
                   />
                 </Th>
                 <Th textAlign="center" fontWeight="bold" fontSize="1rem">
-                  생년
+                  관리자생성일
                   <IconButton
-                    icon={renderSortIcon('birthYear')}
-                    onClick={() => requestSort('birthYear')}
-                    aria-label="Sort 생년"
+                    icon={renderSortIcon('createdAt')}
+                    onClick={() => requestSort('createdAt')}
+                    aria-label="Sort 관리자생성일"
                     size="xs"
                     ml={2}
                   />
                 </Th>
                 <Th textAlign="center" fontWeight="bold" fontSize="1rem">
-                  나이
+                  관리자수정일
                   <IconButton
-                    icon={renderSortIcon('age')}
-                    onClick={() => requestSort('age')}
-                    aria-label="Sort 나이"
+                    icon={renderSortIcon('updatedAt')}
+                    onClick={() => requestSort('updatedAt')}
+                    aria-label="Sort 마지막관리자수정일"
                     size="xs"
                     ml={2}
                   />
                 </Th>
                 <Th textAlign="center" fontWeight="bold" fontSize="1rem">
-                  성별
+                  관리자상태
                   <IconButton
-                    icon={renderSortIcon('gender')}
-                    onClick={() => requestSort('gender')}
-                    aria-label="Sort 성별"
+                    icon={renderSortIcon('adminActive')}
+                    onClick={() => requestSort('adminActive')}
+                    aria-label="Sort 관리자상태"
                     size="xs"
                     ml={2}
                   />
                 </Th>
-                <Th textAlign="center" fontWeight="bold" fontSize="1rem">
-                  위치
-                  <IconButton
-                    icon={renderSortIcon('location')}
-                    onClick={() => requestSort('location')}
-                    aria-label="Sort 위치"
-                    size="xs"
-                    ml={2}
-                  />
-                </Th>
-                <Th textAlign="center" fontWeight="bold" fontSize="1rem">
-                  상태
-                  <IconButton
-                    icon={renderSortIcon('active')}
-                    onClick={() => requestSort('active')}
-                    aria-label="Sort 상태"
-                    size="xs"
-                    ml={2}
-                  />
-                </Th>
+                {/*<Th textAlign="center" fontWeight="bold" fontSize="1rem">*/}
+                {/*  권한*/}
+                {/*  <IconButton*/}
+                {/*    icon={renderSortIcon('권한')}*/}
+                {/*    onClick={() => requestSort('권한')}*/}
+                {/*    aria-label="Sort 권한"*/}
+                {/*    size="xs"*/}
+                {/*    ml={2}*/}
+                {/*  />*/}
+                {/*</Th>*/}
                 <Th textAlign="center">
                   <Button
                     colorScheme="red"
@@ -299,27 +275,42 @@ const UserSetting: React.FC = () => {
               </Tr>
             </Thead>
             <Tbody>
-              {data.map(row => (
-                <Tr key={row.id}>
+              {sortedData.map(row => (
+                <Tr key={row.adminId}>
                   <Td textAlign="center">
                     <Checkbox
-                      isChecked={selectedRows.includes(row.id)}
-                      onChange={() => handleSelectRow(row.id)}
+                      isChecked={selectedRows.includes(row.adminId)}
+                      onChange={() => handleSelectRow(row.adminId)}
                     />
                   </Td>
-                  <Td textAlign="center">{row.id}</Td>
-                  <Td textAlign="center">{row.nickname}</Td>
-                  <Td textAlign="center">{row.email}</Td>
-                  <Td textAlign="center">{row.birthYear}</Td>
-                  <Td textAlign="center">{row.age}</Td>
-                  <Td textAlign="center">{row.gender}</Td>
-                  <Td textAlign="center">{row.location}</Td>
-                  <Td textAlign="center">{row.active}</Td>
+                  <Td textAlign="center">{row.adminId}</Td>
+                  {/*<Td textAlign="center">{row.관리자명}</Td>*/}
+                  <Td textAlign="center">{row.adminEmail}</Td>
+                  <Td textAlign="center">{row.createdAt}</Td>
+                  <Td textAlign="center">{row.updatedAt}</Td>
+                  <Td textAlign="center">{row.adminActive}</Td>
+                  {/*<Td textAlign="center">{row.권한}</Td>*/}
+                  {/*<Td textAlign="center">*/}
+                  {/*<Button*/}
+                  {/*  colorScheme="blue"*/}
+                  {/*  onClick={() =>*/}
+                  {/*    changePermission(*/}
+                  {/*      row.adminId,*/}
+                  {/*      row.권한 === '열람' ? '편집' : '열람',*/}
+                  {/*    )*/}
+                  {/*  }*/}
+                  {/*  isDisabled={!currentUserCanChangePermissions}*/}
+                  {/*>*/}
+                  {/*  권한변경*/}
+                  {/*</Button>*/}
+                  {/*</Td>*/}
                   <Td textAlign="center">
                     <Button
                       colorScheme="red"
-                      onClick={() => handleDelete(row.id)}
-                      isDisabled={row.active === 'inactive'}
+                      onClick={() => handleDelete(row.adminId)}
+                      // isDisabled={
+                      //   row.adminActive === 'active' || row.권한 === '관리자'
+                      // }
                     >
                       삭제
                     </Button>
@@ -347,4 +338,4 @@ const UserSetting: React.FC = () => {
   );
 };
 
-export default UserSetting;
+export default AdminSetting;
