+++ src/pages/GameStat.tsx
